+++ lib/data/datasources/firebase_services/area_firebase_service.dart
@@ -2,13 +2,13 @@ import 'dart:async';
 import 'dart:developer';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:stepper/data/model2/models2.dart';
+import 'package:stepper/data/models/models.dart';
 import 'package:stepper/data/helpers/firestore_helpers.dart';
 
-class AreaFirebaseService2 {
+class AreaFirebaseService {
   final FirebaseFirestore firestore;
 
-  AreaFirebaseService2({required this.firestore});
+  AreaFirebaseService({required this.firestore});
 
   Future<List<Area>> getAllAreas() async {
     final querySnapshot = await firestore.collection('areas').get();
