+++ lib/data/datasources/firebase_services/post_firebase_service.dart
@@ -1,7 +1,7 @@
 import 'dart:developer';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:stepper/data/model2/models2.dart';
+import 'package:stepper/data/models/models.dart';
 import 'package:stepper/data/helpers/firestore_helpers.dart';
 
 class PostFirebaseService {
