+++ lib/data/datasources/firebase_services/profession_firebase_service.dart
@@ -1,5 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:stepper/data/model2/models2.dart';
+import 'package:stepper/data/models/models.dart';
 
 class ProfessionFirebaseService {
   final FirebaseFirestore firestore;
