+++ lib/data/models/user.dart
@@ -1,8 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stepper/data/model2/helpers/timestamp_converter.dart';
-import 'package:stepper/data/model2/models2.dart';
+import 'package:stepper/data/models/helpers/timestamp_converter.dart';
+import 'package:stepper/data/models/models.dart';
 
 part 'user.g.dart';
